# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "TallerMecanico | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5290/"
    }
  ],
  "paths": {
...
    "/api/v{version}/ordenes-trabajo": {
      "get": {
        "tags": [
          "OrdenesTrabajo"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "OrdenesTrabajo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrdenesTrabajo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v{version}/ordenes-trabajo/{id}": {
      "get": {
        "tags": [
          "OrdenesTrabajo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "OrdenesTrabajo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrdenesTrabajo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "OrdenesTrabajo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v{version}/clientes": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Clientes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cliente"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v{version}/clientes/{id}": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cliente"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Cliente": {
        "type": "object",
        "properties": {
          "clienteId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nombre": {
            "type": "string"
          },
          "apellido": {
            "type": "string"
          },
          "telefono": {
            "type": [
              "null",
              "string"
            ]
          },
          "correo": {
            "type": [
              "null",
              "string"
            ]
          },
          "direccion": {
            "type": [
              "null",
              "string"
            ]
          },
          "ordenesTrabajos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrdenesTrabajo"
            }
          },
          "vehiculos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vehiculo"
            }
          }
        }
      },
      "DetalleRepuesto": {
        "type": "object",
        "properties": {
          "detalleRepuestoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "ordenId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "repuestoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "cantidad": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "precioUnitario": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "orden": {
            "$ref": "#/components/schemas/OrdenesTrabajo"
          },
          "repuesto": {
            "$ref": "#/components/schemas/Repuesto"
          }
        }
      },
      "DetalleServicio": {
        "type": "object",
        "properties": {
          "detalleServicioId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "ordenId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "servicioId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "cantidad": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "precioUnitario": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "orden": {
            "$ref": "#/components/schemas/OrdenesTrabajo"
          },
          "servicio": {
            "$ref": "#/components/schemas/Servicio"
          }
        }
      },
      "Especialidade": {
        "type": "object",
        "properties": {
          "especialidadId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nombre": {
            "type": "string"
          },
          "descripcion": {
            "type": [
              "null",
              "string"
            ]
          },
          "mecanicos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Mecanico"
            }
          }
        }
      },
...
      "Marca": {
        "type": "object",
        "properties": {
          "marcaId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nombre": {
            "type": "string"
          },
          "modelos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Modelo"
            }
          }
        }
      },
      "Mecanico": {
        "type": "object",
        "properties": {
          "mecanicoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nombre": {
            "type": "string"
          },
          "apellido": {
            "type": "string"
          },
          "telefono": {
            "type": [
              "null",
              "string"
            ]
          },
          "especialidadId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "especialidad": {
            "$ref": "#/components/schemas/Especialidade"
          },
          "ordenesTrabajos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrdenesTrabajo"
            }
          }
        }
      },
      "Modelo": {
        "type": "object",
        "properties": {
          "modeloId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nombre": {
            "type": "string"
          },
          "marcaId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "marca": {
            "$ref": "#/components/schemas/Marca"
          },
          "vehiculos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vehiculo"
            }
          }
        }
      },
      "OrdenesTrabajo": {
        "type": "object",
        "properties": {
          "ordenId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "fechaIngreso": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "estado": {
            "type": "string"
          },
          "observaciones": {
            "type": [
              "null",
              "string"
            ]
          },
          "clienteId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "vehiculoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "mecanicoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "cliente": {
            "$ref": "#/components/schemas/Cliente"
          },
          "detalleRepuestos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DetalleRepuesto"
            }
          },
          "detalleServicios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DetalleServicio"
            }
          },
          "mecanico": {
            "$ref": "#/components/schemas/Mecanico"
          },
          "pago": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Pago"
              }
            ]
          },
          "vehiculo": {
            "$ref": "#/components/schemas/Vehiculo"
          }
        }
      },
      "Pago": {
        "type": "object",
        "properties": {
          "pagoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "ordenId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "montoTotal": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "metodoPago": {
            "type": "string"
          },
          "fechaPago": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "estado": {
            "type": "string"
          },
          "orden": {
            "$ref": "#/components/schemas/OrdenesTrabajo"
          }
        }
      },
      "Repuesto": {
        "type": "object",
        "properties": {
          "repuestoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nombre": {
            "type": "string"
          },
          "descripcion": {
            "type": [
              "null",
              "string"
            ]
          },
          "precio": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "stock": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detalleRepuestos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DetalleRepuesto"
            }
          }
        }
      },
...
      "Servicio": {
        "type": "object",
        "properties": {
          "servicioId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nombre": {
            "type": "string"
          },
          "descripcion": {
            "type": [
              "null",
              "string"
            ]
          },
          "costo": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "detalleServicios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DetalleServicio"
            }
          }
        }
      },
...
      "Vehiculo": {
        "type": "object",
        "properties": {
          "vehiculoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "patente": {
            "type": "string"
          },
          "anio": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "color": {
            "type": [
              "null",
              "string"
            ]
          },
          "clienteId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "modeloId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "cliente": {
            "$ref": "#/components/schemas/Cliente"
          },
          "modelo": {
            "$ref": "#/components/schemas/Modelo"
          },
          "ordenesTrabajos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrdenesTrabajo"
            }
          }
        }
      }
    }
  },
...
}